Detach Tickets class to TicketsMixin
